feat: Tooling -  use `.ts` file extensions and `esnext` module resolution. - Rename `start` to `dev` - Replace `userId` with `user` and appropriate Mongoose type to allow `.populate()` usage - Add `Procfile` for heroku deployment - Remove invalid `path` for `vincent-app-dca` (per nx prompt, causing build errors)
