+++ Procfile
@@ -1,0 +1,1 @@
+web: pnpm start-dca-backend
