+++ packages/vincent-service-dca/test/helpers/test-server.ts
@@ -1,4 +1,4 @@
-import { Server } from '../../src/lib/server.mjs';
+import { Server } from '../../src/lib/server';
 
 export class TestServer extends Server {
   constructor() {
