+++ packages/vincent-service-dca/test/routes/user.routes.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach, afterEach } from '@jest/globals';
 import fetch from 'node-fetch';
-import '../setup.mts';
-import { TestServer } from '../helpers/test-server.mjs';
+import '../setup.ts';
+import { TestServer } from '../helpers/test-server';
 
 describe('User Routes', () => {
   let server: TestServer;
